+++ driver/src/main/scala/org/mongodb/scala/model/Aggregates.scala
@@ -30,6 +30,97 @@ import com.mongodb.client.model.{Aggregates => JAggregates}
  */
 object Aggregates {
   /**
+   * Creates an \$addFields pipeline stage
+   *
+   * @param fields the fields to add
+   * @return the \$addFields pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/addFields/ \$addFields]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def addFields(fields: Field[_]*): Bson = JAggregates.addFields(fields.asJava)
+
+  /**
+   * Creates a \$bucket pipeline stage
+   *
+   * @param groupBy    the criteria to group By
+   * @param boundaries the boundaries of the buckets
+   * @tparam TExpression the groupBy expression type
+   * @tparam TBoundary    the boundary type
+   * @return the \$bucket pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/bucket/ \$bucket]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def bucket[TExpression, TBoundary](groupBy: TExpression, boundaries: TBoundary*): Bson =
+    JAggregates.bucket(groupBy, boundaries.asJava)
+
+  /**
+   * Creates a \$bucket pipeline stage
+   *
+   * @param groupBy    the criteria to group By
+   * @param boundaries the boundaries of the buckets
+   * @param options    the optional values for the \$bucket stage
+   * @tparam TExpression the groupBy expression type
+   * @tparam TBoundary    the boundary type
+   * @return the \$bucket pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/bucket/ \$bucket]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def bucket[TExpression, TBoundary](groupBy: TExpression, options: BucketOptions, boundaries: TBoundary*): Bson =
+    JAggregates.bucket(groupBy, boundaries.asJava, options)
+
+  /**
+   * Creates a \$bucketAuto pipeline stage
+   *
+   * @param groupBy    the criteria to group By
+   * @param buckets the number of the buckets
+   * @tparam TExpression the groupBy expression type
+   * @return the \$bucketAuto pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/bucketAuto/ \$bucketAuto]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def bucketAuto[TExpression, TBoundary](groupBy: TExpression, buckets: Int): Bson = JAggregates.bucketAuto(groupBy, buckets)
+
+  /**
+   * Creates a \$bucketAuto pipeline stage
+   *
+   * @param groupBy    the criteria to group By
+   * @param buckets the number of the buckets
+   * @param options the optional values for the \$bucketAuto stage
+   * @tparam TExpression the groupBy expression type
+   * @return the \$bucketAuto pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/bucketAuto/ \$bucketAuto]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def bucketAuto[TExpression, TBoundary](groupBy: TExpression, buckets: Int, options: BucketAutoOptions): Bson =
+    JAggregates.bucketAuto(groupBy, buckets, options)
+
+  /**
+   * Creates a \$count pipeline stage using the field name "count" to store the result
+   *
+   * @return the \$count pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/count/ \$count]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def count(): Bson = JAggregates.count()
+
+  /**
+   * Creates a \$count pipeline stage using the named field to store the result
+   *
+   * @param field the field in which to store the count
+   * @return the \$count pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/count/ \$count]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def count(field: String): Bson = JAggregates.count(field)
+
+  /**
    * Creates a `\$match` pipeline stage for the specified filter
    *
    * @param filter the filter to match
@@ -52,6 +143,53 @@ object Aggregates {
   def filter(filter: Bson): Bson = `match`(filter) //scalastyle:ignore
 
   /**
+   * Creates a \$facet pipeline stage
+   *
+   * @param facets the facets to use
+   * @return the new pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/facet/ \$facet]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def facet(facets: Facet*): Bson = JAggregates.facet(facets.asJava)
+
+  /**
+   * Creates a \$graphLookup pipeline stage for the specified filter
+   *
+   * @param from             the collection to query
+   * @param startWith        the expression to start the graph lookup with
+   * @param connectFromField the from field
+   * @param connectToField   the to field
+   * @param as               name of field in output document
+   * @tparam TExpression the expression type
+   * @return the \$graphLookup pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/graphLookup/ \$graphLookup]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def graphLookup[TExpression](from: String, startWith: TExpression, connectFromField: String, connectToField: String, as: String): Bson =
+    JAggregates.graphLookup(from, startWith, connectFromField, connectToField, as)
+
+  /**
+   * Creates a graphLookup pipeline stage for the specified filter
+   *
+   * @param from             the collection to query
+   * @param startWith        the expression to start the graph lookup with
+   * @param connectFromField the from field
+   * @param connectToField   the to field
+   * @param as               name of field in output document
+   * @param options          optional values for the graphLookup
+   * @tparam TExpression the expression type
+   * @return the \$graphLookup pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/graphLookup/ \$graphLookup]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def graphLookup[TExpression](from: String, startWith: TExpression, connectFromField: String, connectToField: String, as: String,
+                               options: GraphLookupOptions): Bson =
+    JAggregates.graphLookup(from, startWith, connectFromField, connectToField, as, options)
+
+  /**
    * Creates a `\$project` pipeline stage for the specified projection
    *
    * @param projection the projection
@@ -62,6 +200,18 @@ object Aggregates {
   def project(projection: Bson): Bson = JAggregates.project(projection)
 
   /**
+   * Creates a \$replaceRoot pipeline stage
+   *
+   * @param value the new root value
+   * @tparam TExpression the new root type
+   * @return the \$replaceRoot pipeline stage
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/replaceRoot/ \$replaceRoot]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def replaceRoot[TExpression](value: TExpression): Bson = JAggregates.replaceRoot(value)
+
+  /**
    * Creates a `\$sort` pipeline stage for the specified sort specification
    *
    * @param sort the sort specification
@@ -70,6 +220,19 @@ object Aggregates {
    * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/sort/#sort-aggregation \$sort]]
    */
   def sort(sort: Bson): Bson = JAggregates.sort(sort)
+
+  /**
+   * Creates a \$sortByCount pipeline stage for the specified filter
+   *
+   * @param filter the filter specification
+   * @tparam TExpression the expression type
+   * @return the \$sortByCount pipeline stage
+   * @see Sorts
+   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/sortByCount \$sortByCount]]
+   * @since 1.2
+   * @note Requires MongoDB 3.4 or greater
+   */
+  def sortByCount[TExpression](filter: TExpression): Bson = JAggregates.sortByCount(filter)
 
   /**
    * Creates a `\$skip` pipeline stage
@@ -102,7 +265,6 @@ object Aggregates {
   /**
    * Creates a `\$lookup` pipeline stage for the specified filter
    *
-   * '''Note:'''Requires MongoDB 3.2 or greater
    * @param from the name of the collection in the same database to perform the join with.
    * @param localField specifies the field from the local collection to match values against.
    * @param foreignField specifies the field in the from collection to match values against.
@@ -110,6 +272,7 @@ object Aggregates {
    * @return the `\$lookup` pipeline stage
    * @since 1.1
    * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/lookup/ \$lookup]]
+   * @note Requires MongoDB 3.2 or greater
    */
   def lookup(from: String, localField: String, foreignField: String, as: String): Bson =
     JAggregates.lookup(from, localField, foreignField, as)