+++ docs/reference/content/changelog.md
@@ -10,7 +10,8 @@ title = "Changelog"
 
 Changes between released versions
 
-### 1.2.0-SNAPSHOT
+### 1.2.0
+  * Added support for the new Aggregation stages in 3.4 [SCALA-258](https://jira.mongodb.org/browse/SCALA-258)
   * Added Collation support [SCALA-249](https://jira.mongodb.org/browse/SCALA-249)
   * Added support for BsonDecimal128 [SCALA-241](https://jira.mongodb.org/browse/SCALA-241)
   * Added support for ReadConcern.LINEARIZABLE [SCALA-247](https://jira.mongodb.org/browse/SCALA-247)